+++ tools/PSc8y/Public-manual/Watch-Event.ps1
@@ -85,16 +85,11 @@ Watch all events for a device
             continue
         }
 
-        $c8y = Get-ClientBinary
-
-        $c8yargs = New-Object System.Collections.ArrayList
-        $null = $c8yargs.AddRange(@("events", "subscribe"))
-        $Parameters.Keys | ForEach-Object {
-            $null = $c8yargs.AddRange(@("$_", $Parameters[$_]))
-        }
-
-        Invoke-BinaryProcess $c8y -RedirectOutput -ArgumentList $c8yargs |
-            Add-PowershellType "application/json"
+        Invoke-ClientCommand `
+            -Noun "events" `
+            -Verb "subscribe" `
+            -Parameters $Parameters `
+            -Type "application/json"
     }
 
     End {}