+++ tools/PSc8y/Public-manual/Watch-NotificationChannels.ps1
@@ -88,14 +88,11 @@ Watch all types of notifications for a device for 90 seconds
             continue
         }
 
-        $c8yargs = New-Object System.Collections.ArrayList
-        $null = $c8yargs.AddRange(@("realtime", "subscribeAll"))
-        $Parameters.Keys | ForEach-Object {
-            $null = $c8yargs.AddRange(@("$_", $Parameters[$_]))
-        }
-
-        Invoke-BinaryProcess (Get-ClientBinary) -RedirectOutput -ArgumentList $c8yargs |
-            Add-PowershellType "application/json"
+        Invoke-ClientCommand `
+            -Noun "realtime" `
+            -Verb "subscribeAll" `
+            -Parameters $Parameters `
+            -Type "application/json"
     }
 
     End {}